Criando bomba remota

+++ src/main/java/br/unb/unbomber/systems/TimeSystem.java
@@ -40,24 +40,27 @@ public class TimeSystem extends BaseSystem {
 
 		// list of
 		List<Timer> toRemoveList = new ArrayList<Timer>();
-		for (Component component : timedEffects) {
-			Timer timeEffect = (Timer) component;
-			timeEffect.tick();
-			/*
-			 * Add a Event when the the time is up
-			 */
-			if (timeEffect.isOver()) {
-				getEntityManager().addEvent(timeEffect.getEvent());
+		if(timedEffects != null){
+			for (Component component : timedEffects) {
+				Timer timeEffect = (Timer) component;
+				timeEffect.tick();
+				/*
+				 * Add a Event when the the time is up
+				 */
+				if (timeEffect.isOver()) {
+					getEntityManager().addEvent(timeEffect.getEvent());
 
-				// add to a list of components to remove, as we can't remove
-				// because we can't change a list when we are iterating in it.
-				toRemoveList.add(timeEffect);
+					// add to a list of components to remove, as we can't remove
+					// because we can't change a list when we are iterating in it.
+					toRemoveList.add(timeEffect);
+				}
+			}
+
+			for (Component component : toRemoveList) {
+				getEntityManager().remove(component);
 			}
 		}
-
-		for (Component component : toRemoveList) {
-			getEntityManager().remove(component);
-		}
+		
 
 	}
